+++ server/auth/index.js
@@ -1,31 +1,2 @@
-import jwt from "jsonwebtoken";
-import dotenv from "dotenv";
-
-dotenv.config();
-
-const APP_SECRET = process.env.APP_SECRET;
-
-function getTokenPayload(token) {
-  return jwt.verify(token, APP_SECRET);
-}
-
-function getUserId(req, authToken) {
-  if (req) {
-    const authHeader = req.headers.authorization;
-    if (authHeader) {
-      const token = authHeader.replace("Bearer ", "");
-      if (!token) {
-        throw new Error("No token found");
-      }
-      const { userId } = getTokenPayload(token);
-      return userId;
-    }
-  } else if (authToken) {
-    const { userId } = getTokenPayload(authToken);
-    return userId;
-  }
-
-  throw new Error("Not authenticated");
-}
-
-export { getTokenPayload, getUserId };
+export { getTokenPayload, getUserId } from "./utils";
+export { getUserDetails, signup, login, testAuthentication } from "./resolvers";
