+++ server/auth/resolvers/index.js
@@ -1,0 +1,4 @@
+export { getUserDetails } from "./getUserDetails";
+export { login } from "./login";
+export { signup } from "./signup";
+export { testAuthentication } from "./testAuthentication";
