+++ server/auth/resolvers/login.js
@@ -1,39 +1,9 @@
-import {
-  insertUser,
-  findUserByEmail,
-  findUserById,
-} from "../database/database";
+import { findUserByEmail } from "../database/database";
 import dotenv from "dotenv";
 import bcrypt from "bcrypt";
 import jwt from "jsonwebtoken";
 
 dotenv.config();
-
-async function signup(parent, args, context, info) {
-  const password = await bcrypt.hash(args.password, 10);
-  try {
-    const userRes = await insertUser(args.name, args.email, password);
-    const user = userRes[0];
-    const token = jwt.sign({ userId: user._id }, process.env.APP_SECRET);
-    return {
-      result: {
-        token,
-        user,
-      },
-      error: null,
-    };
-  } catch (error) {
-    if (error.message === "EXISTING_USER") {
-      return {
-        result: null,
-        error: {
-          name: "EXISTING_USER",
-          message: "User already exists, please try a different email",
-        },
-      };
-    }
-  }
-}
 
 async function login(parent, args, context, info) {
   try {
@@ -79,20 +49,4 @@ async function login(parent, args, context, info) {
   }
 }
 
-async function testAuthentication(parent, args, context, info) {
-  const { userId } = context;
-  return userId;
-}
-
-async function getUserDetails(parent, args, context, info) {
-  const { userId } = context;
-  try {
-    const user = await findUserById(userId);
-    return user;
-  } catch (error) {
-    console.log(error);
-    return null;
-  }
-}
-
-export { signup, login, testAuthentication, getUserDetails };
+export { login };
