+++ server/auth/resolvers/testAuthentication.js
@@ -1,0 +1,6 @@
+async function testAuthentication(parent, args, context, info) {
+  const { userId } = context;
+  return userId;
+}
+
+export { testAuthentication };
