+++ server/database/client.js
@@ -1,0 +1,10 @@
+import { MongoClient } from "mongodb";
+import dotenv from "dotenv";
+
+dotenv.config();
+
+const uri = process.env.MONGODB_URL;
+
+const client = new MongoClient(uri);
+
+export { client };
