+++ server/database/index.js
@@ -1,0 +1,5 @@
+export { client } from "./client";
+export { findUserByEmail } from "./findUserByEmail";
+export { findUserById } from "./findUserById";
+export { insertUser } from "./insertUser";
+export { testConnection } from "./testConnection";
